Refacror
