+++ packages/devextreme/js/common/charts.d.ts
@@ -2,6 +2,7 @@ import { HorizontalAlignment } from '../common';
 import { Format } from '../localization';
 import { BaseLegendItem } from '../viz/common';
 import { baseSeriesObject } from '../viz/chart';
+import { RepeatUnion } from '../core';
 
 /**
  * @public
@@ -45,30 +46,12 @@ export type ChartsLabelOverlap = 'hide' | 'none' | 'stack';
  */
 export type ChartsDataType = 'datetime' | 'numeric' | 'string';
 
-type BuildTuple<Length extends number, Result extends unknown[] = []> = Result['length'] extends Length
-  ? Result
-  : BuildTuple<Length, [...Result, unknown]>;
-
-type Tail<T extends unknown[]> = T extends [unknown, ...infer Rest] ? Rest : [];
-
-type RepeatUnionInternal<
-  T extends string,
-  DepthTuple extends unknown[],
-> = DepthTuple['length'] extends 0
-    ? never
-    : T | `${T}${RepeatUnionInternal<T, Tail<DepthTuple>>}`;
-
-type RepeatUnion<
-  T extends string,
-  Depth extends number,
-> = RepeatUnionInternal<T, BuildTuple<Depth>>;
-
 /**
  * @public
  * @namespace DevExpress.common.charts
  * @type 'dash'|'dot'|'longDash'|'solid'
  */
-export type DashStyle = RepeatUnion<'dash' | 'dot' | 'longDash' | 'solid', 7>;
+export type DashStyle = RepeatUnion<'dash' | 'dot' | 'longDash' | 'solid', 5>;
 
 /**
  * @public
